Homework 3, added ViewModel for EditorialFragment - 27.09.2021
