+++ app/src/main/java/com/example/kts_android_09_2021/fragments/login_fragment/StateFields.kt
@@ -1,0 +1,6 @@
+package com.example.kts_android_09_2021.fragments.login_fragment
+
+data class StateFields(
+    var emailIsCorrect: Boolean,
+    var passwordIsCorrect: Boolean
+)
